backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Audio
python-vlc==3.0.20123
mutagen==1.47.0

# AI
anthropic==0.18.1

# Google APIs
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Notifications
twilio==8.13.0
pywebpush==1.14.0

# Background Tasks
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
